# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Scraping Service",
    "description": "URL scraping service with multi-provider support (Scrape.do default, Firecrawl) - extracts company information from websites",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3010"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "ScrapingProvider": {
        "type": "string",
        "enum": [
          "scrape-do",
          "firecrawl"
        ]
      },
      "ScrapeResult": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "industry": {
            "type": "string",
            "nullable": true
          },
          "employeeCount": {
            "type": "string",
            "nullable": true
          },
          "foundedYear": {
            "type": "number",
            "nullable": true
          },
          "headquarters": {
            "type": "string",
            "nullable": true
          },
          "website": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "linkedinUrl": {
            "type": "string",
            "nullable": true
          },
          "twitterUrl": {
            "type": "string",
            "nullable": true
          },
          "products": {
            "nullable": true
          },
          "services": {
            "nullable": true
          },
          "rawMarkdown": {
            "type": "string",
            "nullable": true
          },
          "rawHtml": {
            "type": "string",
            "nullable": true,
            "description": "Raw page HTML. Populated only when options.formats includes \"rawHtml\"; otherwise null. Request-scoped (not cached)."
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          }
        },
        "required": [
          "id",
          "url",
          "companyName",
          "description",
          "industry",
          "employeeCount",
          "foundedYear",
          "headquarters",
          "website",
          "email",
          "phone",
          "linkedinUrl",
          "twitterUrl",
          "rawMarkdown",
          "rawHtml",
          "createdAt"
        ]
      },
      "ScrapeResponse": {
        "type": "object",
        "properties": {
          "cached": {
            "type": "boolean"
          },
          "provider": {
            "$ref": "#/components/schemas/ScrapingProvider"
          },
          "requestId": {
            "type": "string"
          },
          "runId": {
            "type": "string"
          },
          "result": {
            "$ref": "#/components/schemas/ScrapeResult"
          }
        },
        "required": [
          "cached",
          "result"
        ]
      },
      "InsufficientCreditsResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "balance_cents": {
            "type": "number",
            "nullable": true
          },
          "required_cents": {
            "type": "number"
          }
        },
        "required": [
          "error",
          "balance_cents",
          "required_cents"
        ]
      },
      "ScrapeErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "ScrapeOptions": {
        "type": "object",
        "properties": {
          "formats": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "markdown",
                "html",
                "rawHtml",
                "links",
                "screenshot"
              ]
            }
          },
          "onlyMainContent": {
            "type": "boolean"
          },
          "includeTags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "excludeTags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "waitFor": {
            "type": "number"
          },
          "timeout": {
            "type": "integer",
            "minimum": 1000,
            "maximum": 300000
          }
        }
      },
      "ScrapeRequest": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "provider": {
            "$ref": "#/components/schemas/ScrapingProvider"
          },
          "sourceService": {
            "type": "string"
          },
          "sourceRefId": {
            "type": "string"
          },
          "skipCache": {
            "type": "boolean",
            "default": false
          },
          "enrich": {
            "type": "boolean",
            "description": "Run company-info enrichment. Defaults to true when omitted (current behavior). Set false to skip enrichment and fetch the raw body cheaply (raw-fetch mode)."
          },
          "render": {
            "type": "boolean",
            "description": "Force scrape.do JS rendering (render=true&super=true). Use for client-rendered pages whose content is not in the initial HTML."
          },
          "options": {
            "$ref": "#/components/schemas/ScrapeOptions"
          },
          "brandIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "campaignId": {
            "type": "string"
          },
          "workflowSlug": {
            "type": "string"
          },
          "featureSlug": {
            "type": "string"
          },
          "audienceId": {
            "type": "string"
          }
        },
        "required": [
          "url"
        ]
      },
      "ScrapeByUrlResponse": {
        "type": "object",
        "properties": {
          "cached": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "expired": {
            "type": "boolean"
          },
          "result": {
            "$ref": "#/components/schemas/ScrapeResult"
          }
        },
        "required": [
          "cached",
          "expired",
          "result"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/scrape": {
      "post": {
        "summary": "Scrape a URL and extract company information (provider: scrape-do | firecrawl)",
        "security": [
          {
            "apiKey": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScrapeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Scrape result (cached or fresh)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScrapeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Insufficient credits (platform key only)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsufficientCreditsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Scrape failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScrapeErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/scrape/{id}": {
      "get": {
        "summary": "Get a scrape result by ID",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Scrape result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/components/schemas/ScrapeResult"
                    }
                  },
                  "required": [
                    "result"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Result not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/scrape/by-url": {
      "get": {
        "summary": "Get cached result by URL",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "required": true,
            "name": "url",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Cached scrape result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScrapeByUrlResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing url query param",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No cached result found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}