# src/routes/scrape.ts
import { Router } from "express";
import { eq, and, gt } from "drizzle-orm";
import { db } from "../db/index.js";
import { scrapeRequests, scrapeResults, scrapeCache } from "../db/schema.js";
import { scrapeUrl, normalizeUrl } from "../lib/firecrawl.js";
import { resolveKey, KeyServiceError } from "../lib/key-client.js";
import { scrapeWithEscalation } from "../lib/scrape-chain.js";
import { createRun, updateRunStatus, addCosts } from "../lib/runs-client.js";
import { authorizeCredits } from "../lib/billing-client.js";
import { AuthenticatedRequest } from "../middleware/auth.js";
import { ScrapeRequestSchema, ScrapingProvider } from "../schemas.js";
import { sanitizeForPostgres, MAX_MARKDOWN_LENGTH, MAX_HTML_LENGTH } from "../lib/sanitize.js";

const DEFAULT_PROVIDER: ScrapingProvider = "scrape-do";

/** Map provider name to key-service provider identifier */
const PROVIDER_KEY_NAME: Record<ScrapingProvider, string> = {
  "scrape-do": "scrape-do",
  firecrawl: "firecrawl",
};

/** Map provider name to cost name for billing/runs */
const PROVIDER_COST_NAME: Record<ScrapingProvider, string> = {
  "scrape-do": "scrape-do-credit",
  firecrawl: "firecrawl-scrape-credit",
};

const router = Router();

// Cache duration: 7 days
const CACHE_DURATION_MS = 7 * 24 * 60 * 60 * 1000;

/**
 * POST /scrape
 * Scrape a URL and extract company information
 */
router.post("/scrape", async (req: AuthenticatedRequest, res) => {
  let runId: string | undefined;
  try {
    const parsed = ScrapeRequestSchema.safeParse(req.body);

    if (!parsed.success) {
      return res
        .status(400)
        .json({ error: "Invalid request", details: parsed.error.flatten() });
    }

    const {
      url,
      provider: requestedProvider,
      sourceService,
      sourceRefId,
      options,
      skipCache,
      enrich,
      render,
      brandIds,
      campaignId,
      workflowSlug,
      featureSlug,
      audienceId,
    } = parsed.data;

    const provider: ScrapingProvider = requestedProvider ?? DEFAULT_PROVIDER;

    // Raw-fetch mode: caller wants the raw page body cheaply, WITHOUT company-info
    // enrichment and WITHOUT polluting the shared company-info cache/result store.
    // Triggered by enrich:false OR a rawHtml format request.
    const wantsRawHtml = options?.formats?.includes("rawHtml") ?? false;
    const rawFetchMode = enrich === false || wantsRawHtml;

    const orgId = (req as AuthenticatedRequest).orgId!;
    const userId = (req as AuthenticatedRequest).userId!;
    const parentRunId = (req as AuthenticatedRequest).runId;

    // Headers take precedence over body fields for tracking
    const effectiveCampaignId = (req as AuthenticatedRequest).campaignId || campaignId;
    const effectiveBrandIds = (req as AuthenticatedRequest).brandIds || brandIds;
    const effectiveWorkflowSlug = (req as AuthenticatedRequest).workflowSlug || workflowSlug;
    const effectiveFeatureSlug = (req as AuthenticatedRequest).featureSlug || featureSlug;
    const effectiveAudienceId = (req as AuthenticatedRequest).audienceId || audienceId;

    const normalized = normalizeUrl(url);

    // Check cache first (unless skipCache is true).
    // Raw-fetch mode always scrapes fresh: rawHtml is not persisted, and we must
    // not serve a markdown-only cached row when raw HTML was requested.
    if (!skipCache && !rawFetchMode) {
      const cached = await db.query.scrapeCache.findFirst({
        where: and(
          eq(scrapeCache.normalizedUrl, normalized),
          eq(scrapeCache.isValid, true),
          gt(scrapeCache.expiresAt, new Date())
        ),
      });

      if (cached) {
        // Get full result
        const result = await db.query.scrapeResults.findFirst({
          where: eq(scrapeResults.id, cached.resultId),
        });

        if (result) {
          return res.json({
            cached: true,
            result: formatResult(result),
          });
        }
      }
    }

    // Resolve provider API key via key-service (auto-resolves org/platform source)
    let providerApiKey: string;
    let keySource: "org" | "platform";
    try {
      const decrypted = await resolveKey({
        provider: PROVIDER_KEY_NAME[provider],
        orgId,
        userId,
        runId: parentRunId,
        campaignId: effectiveCampaignId,
        brandIds: effectiveBrandIds,
        workflowSlug: effectiveWorkflowSlug,
        featureSlug: effectiveFeatureSlug,
        audienceId: effectiveAudienceId,
        caller: { method: "POST", path: "/scrape" },
      });
      providerApiKey = decrypted.key;
      keySource = decrypted.keySource;
    } catch (err) {
      if (err instanceof KeyServiceError) {
        const status = err.statusCode === 404 ? 400 : 502;
        const message =
          err.statusCode === 404
            ? `${provider} API key not configured`
            : `Failed to retrieve ${provider} API key`;
        return res.status(status).json({ error: message });
      }
      throw err;
    }

    // Authorize credits with billing-service (platform keys only)
    // For scrape-do, authorize worst-case 25 credits (render+super) since the chain may escalate.
    // Failed attempts are free — only the successful level is billed.
    const authCostName = PROVIDER_COST_NAME[provider];
    const authQuantity = provider === "scrape-do" ? 25 : 1;
    if (keySource === "platform") {
      try {
        const billingIdentity = { orgId, userId, runId: parentRunId, campaignId: effectiveCampaignId, brandIds: effectiveBrandIds, workflowSlug: effectiveWorkflowSlug, featureSlug: effectiveFeatureSlug, audienceId: effectiveAudienceId};
        const auth = await authorizeCredits(
          [{ costName: authCostName, quantity: authQuantity }],
          authCostName,
          billingIdentity
        );
        if (!auth.sufficient) {
          return res.status(402).json({
            error: "Insufficient credits",
            balance_cents: auth.balance_cents,
            required_cents: auth.required_cents,
          });
        }
      } catch (err) {
        console.error("Billing authorization failed:", err);
        return res.status(502).json({ error: "Billing authorization unavailable" });
      }
    }

    // Create run in RunsService
    // x-run-id = parentRunId so runs-service sets it as the parent
    try {
      const run = await createRun(
        { taskName: "scrape", brandIds: effectiveBrandIds, campaignId: effectiveCampaignId, workflowSlug: effectiveWorkflowSlug, featureSlug: effectiveFeatureSlug, audienceId: effectiveAudienceId},
        { orgId, userId, runId: parentRunId, campaignId: effectiveCampaignId, brandIds: effectiveBrandIds, workflowSlug: effectiveWorkflowSlug, featureSlug: effectiveFeatureSlug, audienceId: effectiveAudienceId}
      );
      runId = run.id;
    } catch (err) {
      console.error("Failed to create run:", err);
    }

    // Create scrape request record
    const [request] = await db
      .insert(scrapeRequests)
      .values({
        sourceService: sourceService || req.sourceService || "unknown",
        orgId,
        sourceRefId,
        runId,
        campaignId: effectiveCampaignId,
        brandIds: effectiveBrandIds,
        workflowSlug: effectiveWorkflowSlug,
        featureSlug: effectiveFeatureSlug,
        audienceId: effectiveAudienceId,
        url,
        provider,
        options: options as any,
        status: "processing",
      })
      .returning();

    // Scrape the URL using the selected provider
    // When provider is scrape-do (default), use the escalation chain:
    // basic → render → render+super �� firecrawl fallback
    let scrapeResponse: { success: boolean; markdown?: string; html?: string; metadata?: any; error?: string };
    let actualCostName: string;
    let actualKeySource: "org" | "platform" = keySource;
    let actualProvider: ScrapingProvider = provider;
    let actualRequestCost = 1;

    if (provider === "firecrawl") {
      scrapeResponse = await scrapeUrl(url, providerApiKey, options || {});
      actualCostName = PROVIDER_COST_NAME["firecrawl"];
    } else {
      const chainResult = await scrapeWithEscalation(
        {
          url,
          scrapeDoApiKey: providerApiKey,
          options: options || {},
          forceRender: render === true,
          resolveFirecrawlKey: async () => {
            const decrypted = await resolveKey({
              provider: "firecrawl",
              orgId,
              userId,
              runId: parentRunId,
              campaignId: effectiveCampaignId,
              brandIds: effectiveBrandIds,
              workflowSlug: effectiveWorkflowSlug,
              featureSlug: effectiveFeatureSlug,
              caller: { method: "POST", path: "/scrape" },
            });
            return { key: decrypted.key, keySource: decrypted.keySource };
          },
        },
        keySource
      );
      scrapeResponse = chainResult.response;
      actualCostName = chainResult.costName;
      actualKeySource = chainResult.keySource;
      actualProvider = chainResult.provider as ScrapingProvider;
      if (chainResult.requestCost != null) {
        actualRequestCost = chainResult.requestCost;
      }
    }

    if (!scrapeResponse.success) {
      // Update request as failed
      await db
        .update(scrapeRequests)
        .set({
          status: "failed",
          errorMessage: scrapeResponse.error,
          completedAt: new Date(),
        })
        .where(eq(scrapeRequests.id, request.id));

      if (runId) {
        updateRunStatus(runId, "failed", { orgId, userId, runId, campaignId: effectiveCampaignId, brandIds: effectiveBrandIds, workflowSlug: effectiveWorkflowSlug, featureSlug: effectiveFeatureSlug, audienceId: effectiveAudienceId}).catch((err) =>
          console.error("Failed to update run status:", err)
        );
      }

      return res.status(500).json({
        error: scrapeResponse.error || "Scrape failed",
        requestId: request.id,
        runId,
      });
    }

    // Raw-fetch mode: return the raw body cheaply, skip company-info enrichment,
    // and do NOT touch the shared company-info cache/result store. scrape.do cost
    // is still declared on the forwarded run (zero-direct-spend invariant preserved).
    if (rawFetchMode) {
      const rawHtml = sanitizeForPostgres(scrapeResponse.html, MAX_HTML_LENGTH);
      const rawMarkdown = sanitizeForPostgres(scrapeResponse.markdown, MAX_MARKDOWN_LENGTH);
      const rawCompanyInfo =
        enrich === false
          ? { companyName: null, description: null, industry: null }
          : extractCompanyInfo(scrapeResponse);

      await db
        .update(scrapeRequests)
        .set({ status: "completed", provider: actualProvider, completedAt: new Date() })
        .where(eq(scrapeRequests.id, request.id));

      if (runId) {
        const runIdentity = { orgId, userId, runId, campaignId: effectiveCampaignId, brandIds: effectiveBrandIds, workflowSlug: effectiveWorkflowSlug, featureSlug: effectiveFeatureSlug, audienceId: effectiveAudienceId};
        Promise.all([
          addCosts(runId, [{ costName: actualCostName, quantity: actualRequestCost, costSource: actualKeySource }], runIdentity),
          updateRunStatus(runId, "completed", runIdentity),
        ]).catch((err) => console.error("[scraping-service] Failed to finalize run:", err));
      }

      return res.json({
        cached: false,
        provider: actualProvider,
        requestId: request.id,
        runId,
        result: formatResult(
          {
            id: request.id,
            url,
            companyName: rawCompanyInfo.companyName,
            description: rawCompanyInfo.description,
            industry: rawCompanyInfo.industry,
            website: url,
            rawMarkdown,
            createdAt: new Date(),
          },
          rawHtml
        ),
      });
    }

    // Extract company info from metadata and markdown
    const companyInfo = extractCompanyInfo(scrapeResponse);

    // Store result
    const expiresAt = new Date(Date.now() + CACHE_DURATION_MS);

    const sanitizedMarkdown = sanitizeForPostgres(scrapeResponse.markdown, MAX_MARKDOWN_LENGTH);
    const sanitizedDescription = sanitizeForPostgres(companyInfo.description);
    const sanitizedCompanyName = sanitizeForPostgres(companyInfo.companyName);

    const [result] = await db
      .insert(scrapeResults)
      .values({
        requestId: request.id,
        url,
        normalizedUrl: normalized,
        companyName: sanitizedCompanyName,
        description: sanitizedDescription,
        industry: companyInfo.industry,
        website: url,
        rawMarkdown: sanitizedMarkdown,
        rawMetadata: scrapeResponse.metadata as any,
        expiresAt,
      })
      .onConflictDoUpdate({
        target: scrapeResults.normalizedUrl,
        set: {
          requestId: request.id,
          url,
          companyName: sanitizedCompanyName,
          description: sanitizedDescription,
          industry: companyInfo.industry,
          website: url,
          rawMarkdown: sanitizedMarkdown,
          rawMetadata: scrapeResponse.metadata as any,
          expiresAt,
          updatedAt: new Date(),
        },
      })
      .returning();

    // Update cache
    await db
      .insert(scrapeCache)
      .values({
        normalizedUrl: normalized,
        resultId: result.id,
        companyName: sanitizedCompanyName,
        industry: companyInfo.industry,
        isValid: true,
        expiresAt,
      })
      .onConflictDoUpdate({
        target: scrapeCache.normalizedUrl,
        set: {
          resultId: result.id,
          companyName: sanitizedCompanyName,
          industry: companyInfo.industry,
          isValid: true,
          expiresAt,
          updatedAt: new Date(),
        },
      });

    // Update request as completed with actual provider that succeeded
    await db
      .update(scrapeRequests)
      .set({
        status: "completed",
        provider: actualProvider,
        completedAt: new Date(),
      })
      .where(eq(scrapeRequests.id, request.id));

    // Report costs and complete run (fire-and-forget)
    if (runId) {
      const runIdentity = { orgId, userId, runId, campaignId: effectiveCampaignId, brandIds: effectiveBrandIds, workflowSlug: effectiveWorkflowSlug, featureSlug: effectiveFeatureSlug, audienceId: effectiveAudienceId};
      Promise.all([
        addCosts(runId, [{ costName: actualCostName, quantity: actualRequestCost, costSource: actualKeySource }], runIdentity),
        updateRunStatus(runId, "completed", runIdentity),
      ]).catch((err) => console.error("Failed to finalize run:", err));
    }

    res.json({
      cached: false,
      provider: actualProvider,
      requestId: request.id,
      runId,
      result: formatResult(result),
    });
  } catch (error: any) {
    console.error("[scraping-service] Scrape error:", error);

    // Close the run so it doesn't stay stuck in "running" forever
    if (runId) {
      const orgId = (req as AuthenticatedRequest).orgId!;
      const userId = (req as AuthenticatedRequest).userId!;
      updateRunStatus(runId, "failed", { orgId, userId, runId }).catch((err) =>
        console.error("[scraping-service] Failed to close run in outer catch:", err)
      );
    }

    // Update request status to "failed" so it doesn't stay stuck in "processing"
    try {
      const parsed = ScrapeRequestSchema.safeParse(req.body);
      if (parsed.success) {
        const normalized = normalizeUrl(parsed.data.url);
        await db
          .update(scrapeRequests)
          .set({
            status: "failed",
            errorMessage: error.message || "Internal server error",
            completedAt: new Date(),
          })
          .where(
            and(
              eq(scrapeRequests.url, parsed.data.url),
              eq(scrapeRequests.status, "processing")
            )
          );
      }
    } catch (updateErr) {
      console.error("[scraping-service] Failed to update request status:", updateErr);
    }

    res.status(500).json({ error: error.message || "Internal server error" });
  }
});

/**
 * GET /scrape/by-url
 * Get cached result by URL
 * NOTE: Must be registered BEFORE /scrape/:id to avoid "by-url" matching as a UUID param
 */
router.get("/scrape/by-url", async (req, res) => {
  try {
    const url = req.query.url as string;

    if (!url) {
      return res.status(400).json({ error: "url query param is required" });
    }

    const normalized = normalizeUrl(url);

    const cached = await db.query.scrapeCache.findFirst({
      where: and(
        eq(scrapeCache.normalizedUrl, normalized),
        eq(scrapeCache.isValid, true)
      ),
    });

    if (!cached) {
      return res.status(404).json({ error: "No cached result found" });
    }

    const result = await db.query.scrapeResults.findFirst({
      where: eq(scrapeResults.id, cached.resultId),
    });

    if (!result) {
      return res.status(404).json({ error: "Result not found" });
    }

    res.json({
      cached: true,
      expired: cached.expiresAt < new Date(),
      result: formatResult(result),
    });
  } catch (error: any) {
    console.error("Get by URL error:", error);
    res.status(500).json({ error: error.message || "Internal server error" });
  }
});

/**
 * GET /scrape/:id
 * Get a scrape result by ID
 */
router.get("/scrape/:id", async (req, res) => {
  try {
    const { id } = req.params;

    const result = await db.query.scrapeResults.findFirst({
      where: eq(scrapeResults.id, id),
    });

    if (!result) {
      return res.status(404).json({ error: "Result not found" });
    }

    res.json({ result: formatResult(result) });
  } catch (error: any) {
    console.error("Get result error:", error);
    res.status(500).json({ error: error.message || "Internal server error" });
  }
});

/**
 * Extract company info from scrape response
 */
function extractCompanyInfo(response: { markdown?: string; metadata?: any }) {
  const metadata = response.metadata || {};

  return {
    companyName: metadata.ogTitle || metadata.title || null,
    description: metadata.ogDescription || metadata.description || null,
    industry: null, // Would need AI extraction for this
  };
}

/**
 * Format result for API response.
 * `rawHtml` is request-scoped (never persisted): the fresh-scrape path passes the
 * scraped body; cache-hit / by-url / by-id read paths leave it null.
 */
function formatResult(result: any, rawHtml: string | null = null) {
  return {
    id: result.id,
    url: result.url,
    companyName: result.companyName ?? null,
    description: result.description ?? null,
    industry: result.industry ?? null,
    employeeCount: result.employeeCount ?? null,
    foundedYear: result.foundedYear ?? null,
    headquarters: result.headquarters ?? null,
    website: result.website ?? null,
    email: result.email ?? null,
    phone: result.phone ?? null,
    linkedinUrl: result.linkedinUrl ?? null,
    twitterUrl: result.twitterUrl ?? null,
    products: result.products ?? null,
    services: result.services ?? null,
    rawMarkdown: result.rawMarkdown ?? null,
    rawHtml,
    createdAt: result.createdAt,
  };
}

export default router;